Refactor and reduce dependencies
