+++ app/src/test/java/dev/arkbuilders/navigator/data/utils/FileUtilsTest.kt
@@ -1,0 +1,5 @@
+package dev.arkbuilders.navigator.data.utils
+
+import org.junit.jupiter.api.Assertions.*
+
+class FileUtilsTest
